+++ src/main/java/racingcar/domain/car/Car.java
@@ -1,6 +1,6 @@
-package racingcar.car;
+package racingcar.domain.car;
 
-import racingcar.random.RandNum;
+import racingcar.domain.random.RandNum;
 
 public class Car {
