+++ src/main/java/racingcar/domain/game/WinnerDecisionByBigLocations.java
@@ -1,8 +1,8 @@
-package racingcar.game;
+package racingcar.domain.game;
 
-import racingcar.car.Car;
-import racingcar.car.Position;
-import racingcar.car.Winners;
+import racingcar.domain.car.Car;
+import racingcar.domain.car.Position;
+import racingcar.domain.car.Winners;
 
 import java.util.ArrayList;
 import java.util.Collections;
