+++ src/test/java/racingcar/domain/car/CarTest.java
@@ -1,4 +1,4 @@
-package racingcar.car;
+package racingcar.domain.car;
 
 import org.assertj.core.api.Assertions;
 import org.assertj.core.api.Condition;
@@ -6,7 +6,9 @@ import org.junit.jupiter.api.DisplayNameGeneration;
 import org.junit.jupiter.api.DisplayNameGenerator.ReplaceUnderscores;
 import org.junit.jupiter.api.Nested;
 import org.junit.jupiter.api.Test;
-import racingcar.random.RandNum;
+import racingcar.domain.car.Car;
+import racingcar.domain.car.Position;
+import racingcar.domain.random.RandNum;
 
 import static org.assertj.core.api.Assertions.assertThat;
