+++ src/test/java/racingcar/domain/random/RandNumTest.java
@@ -1,9 +1,10 @@
-package racingcar.random;
+package racingcar.domain.random;
 
 import org.assertj.core.api.Assertions;
 import org.junit.jupiter.api.DisplayNameGeneration;
 import org.junit.jupiter.api.DisplayNameGenerator;
 import org.junit.jupiter.api.Test;
+import racingcar.domain.random.RandNum;
 
 @DisplayNameGeneration(DisplayNameGenerator.ReplaceUnderscores.class)
 class RandNumTest {
